requirements.txt
api==0.0.7
certifi==2019.11.28
chardet==3.0.4
Django==2.0.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.8
nose==1.3.7
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
urllib3==1.25.7
